Down to memory loss of 576/160 (from 896/288)
I added a check in the FindMostLeft function in AVLCounter.h; there was a problem at the (f->leftSon =NULL) assignation, if f->leftSon wasn't NULL and also wasn't equal to curr, we lost that information (for example, if this function was called from within the RecRemoval function, in which f->leftSon isnt involved, we would have a memory leak). Change is in lines 392-394.